@@ -1,9 +1,9 @@
 #include "OpenGLImpl/Utils/RenderBufferImplUtils.h"
 
-#include "FmtFormat.h"
 #include "GraphicLib/Utilities/RenderBufferUtils.h"
 #include "InternalLogger.h"
 #include <glad/glad.h>
+#include <format>
 #include <string>
 
 
@@ -25,7 +25,7 @@ bool RenderBufferImplUtils::ConvertRenderBufferFormatToFrameBufferAttachment(ERe
             break;
     }
 
-    const std::string& errorText = fmt::format("RenderBufferFormat {}", RenderBufferUtils::RenderBufferFormatToString(format));
+    const std::string& errorText = std::format("RenderBufferFormat {}", RenderBufferUtils::RenderBufferFormatToString(format));
     LOG_INTERNAL_ERROR(errorText.c_str());
     return false;
 }
